add friends to onboarding
